Move service provider and vs extensions to separate classes
